Draft of Celery/Pulse integration
Mozilla has Pulse[1] which is a RabbitMQ service.
We can manage the Pulse account by PulseGuardian[2].

I create a account for testing the Celery integration.
After enable the "BROKER_USE_SSL", the celery worker can connect to the server.

The Pulse user only have the full permissions on "^exchange/<username>/.*$, so I try to to modify the "CELERY_DEFAULT_EXCHANGE", "CELERY_DEFAULT_QUEUE", AND "CELERY_EVENT_QUEUE_PREFIX".

But it still doesn't work :(
The reason is flower cannot access the "celeryev", and celery worker cannot access "reply.celery.pidbox" exchanges, which is controlled by Kombu[3].

It seems like we have to host a RabbitMQ with full permission by our self.

[1] https://wiki.mozilla.org/Auto-tools/Projects/Pulse
[2] https://pulseguardian.mozilla.org/
[3] https://github.com/celery/kombu/blob/master/kombu/pidbox.py#L142-L143